created a submission

+++ script.js
@@ -11,4 +11,5 @@ function range(start,end) {
 // do not change the code below.
 let start = parseInt(prompt("Enter Start of the Range."));
 let end = parseInt(prompt("Enter End Of the Range."));
-alert(range(parseInt(start), parseInt(end)));
+alert(range(parseFloat(start), parseFloat(end)));
+
